requirements.txt
dateparser==1.2.0
httpx==0.27.2
python-telegram-bot==21.4
